requirements.txt
alpaca-trade-api
python-dotenv
pandas
numpy
matplotlib
scikit-learn  # Pour l'IA
textblob
requests
yfinance # conturne les restriction de données de Alpaca
